Done competitors api and slightly changed file structure.
